requirements.txt
astroid==1.5.3
autopep8==1.3.3
backports.csv==1.0.5
certifi==2017.11.5
chardet==3.0.4
cli-helpers==0.2.3
click==6.7
configobj==5.0.6
Django==1.11.7
django-crispy-forms==1.7.0
django-filter==1.1.0
django-money==0.12.1
django-ymaps-admin-widget==1.0.12
djangorestframework==3.7.3
djangorestframework-gis==0.12
docopt==0.6.2
drf-generators==0.3.0
flake8==3.5.0
humanize==0.5.1
idna==2.6
isort==4.2.15
jedi==0.11.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
mimesis==0.0.10
parso==0.1.0
pgcli==1.8.1
pgspecial==1.9.0
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptpython==0.41
py-moneyed==0.7.0
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.7.4
pytz==2017.3
requests==2.18.4
setproctitle==1.1.10
six==1.11.0
sqlparse==0.2.4
terminaltables==3.1.0
urllib3==1.22
wcwidth==0.1.7
wrapt==1.10.11
yapf==0.20.0
